Split the benchmarks that compare against other engines into their own crate
Cargo builds every dev-dependency of a package whenever any of its bench targets
is built, and benches/common.rs referenced every engine unconditionally, so
compiling redb_benchmark from a cold target directory also built RocksDB's
vendored C++ sources: 6m03s of its 7m07s, even though that benchmark only
exercises redb.

A feature flag cannot fix this. Optional dev-dependencies are rejected outright,
and moving rocksdb to a regular dependency pulls bindgen and libloading into
cargo-deny's license check, which prunes build-dependencies reached through a
dev-dependency edge.

Move the benchmarks that compare against other engines, along with their
adapters, to redb-bench-compare, where lmdb, RocksDB, fjall, and SQLite stay
dev-dependencies of a crate that redb-bench does not depend on. The shared
harness becomes redb-bench's library, so it is compiled once rather than once
per benchmark, which also retires the phantom bench target that Cargo
auto-discovered from benches/common.rs.

cargo bench --no-run --bench redb_benchmark now takes 40s instead of 10m13s from
a cold target directory.
