+++ Cargo.toml
@@ -12,7 +12,7 @@ repository.workspace = true
 authors.workspace = true
 
 [workspace]
-members = [".", "crates/redb-bench", "crates/redb-derive", "crates/redb-derive-rename-test", "crates/redb-python"]
+members = [".", "crates/redb-bench", "crates/redb-bench-compare", "crates/redb-derive", "crates/redb-derive-rename-test", "crates/redb-python"]
 default-members = [".", "crates/redb-derive", "crates/redb-derive-rename-test", "crates/redb-python"]
 
 [workspace.package]
