+++ Dockerfile.bench
@@ -12,4 +12,4 @@ RUN curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --de
 
 ADD . /code/redb/
 
-RUN cd /code/redb && cargo bench --no-run -p redb-bench
+RUN cd /code/redb && cargo bench --no-run -p redb-bench -p redb-bench-compare
