+++ crates/redb-bench-compare/benches/syscall_benchmark.rs
@@ -1,7 +1,6 @@
 use tempfile::{NamedTempFile, TempDir};
 
-mod benchmark_dir;
-use benchmark_dir::benchmark_dir;
+use redb_bench::benchmark_dir;
 
 use rand::RngExt;
 use rand::prelude::SliceRandom;
