+++ crates/redb-bench/benches/redb_benchmark.rs
@@ -1,12 +1,9 @@
 use std::{fs, process};
 use tempfile::NamedTempFile;
 
-mod benchmark_dir;
-use benchmark_dir::benchmark_dir;
+use redb_bench::benchmark_dir;
 
-#[expect(dead_code)]
-mod common;
-use common::*;
+use redb_bench::*;
 
 fn main() {
     let _ = env_logger::try_init();
