+++ crates/redb-bench/benches/savepoint_benchmark.rs
@@ -2,8 +2,7 @@
 
 use tempfile::NamedTempFile;
 
-mod benchmark_dir;
-use benchmark_dir::benchmark_dir;
+use redb_bench::benchmark_dir;
 
 use rand::RngExt;
 use redb::{Database, TableDefinition};
